Fix a [-Wshift-negative-value] warning
by assuming a specific type of non-standard behavior was expected
by this code when left-shifting a negative value.

The C standard says that the result of left shifting a negative value
should be undefined, but the version of gcc I'm using, and the C compiler
available at replit.com seem to both agree (according to some testing) that
left shifting negative y by x (i.e. y << x) shall be computed as
(int)(((unsigned)y) << x).  So that's the approach I'm using.

I got inspiration for this fix from this SO answer, which also claims that
99+% of all C compilers take the approach I've chosen:
https://stackoverflow.com/a/29710927

+++ src/gl_main.c
@@ -1900,8 +1900,9 @@ void gld_EndDrawScene(void) {
   gld_Set2DMode();
 
   if (viewangleoffset <= 1024 << ANGLETOFINESHIFT ||
-      viewangleoffset >= -1024
-                             << ANGLETOFINESHIFT) { // don't draw on side views
+      viewangleoffset >=
+          (int)(((unsigned)-1024)
+                << ANGLETOFINESHIFT)) { // don't draw on side views
     R_DrawPlayerSprites();
   }
   if (player->fixedcolormap == 32) {
